chore: Remove clear_namespace script
Since https://github.com/ai-dynamo/dynamo/pull/3350 the MDC is attached
to an etcd lease, so it cleans up on shutdown. We don't need to manually
clear namespace any more.

If there are any remaining non-lease keys (I'm not aware of any, and I
have looked), we will treat that as a bug.

Signed-off-by: Graham King <grahamk@nvidia.com>